Update View and Controller

+++ XGEAR/Models/Product.cs
@@ -14,7 +14,7 @@ namespace XGEAR.Models
     using System.Collections.Generic;
     using System.ComponentModel.DataAnnotations.Schema;
     using System.Web;
-
+    
     public partial class Product
     {
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2214:DoNotCallOverridableMethodsInConstructors")]
@@ -25,15 +25,15 @@ namespace XGEAR.Models
         }
         [NotMapped]
         public HttpPostedFileBase UploadImage { get; set; }
-
+    
         public int ProductID { get; set; }
         public string NamePro { get; set; }
         public string DecriptionPro { get; set; }
         public string Category { get; set; }
-        public Nullable<decimal> Price { get; set; }
+        public Nullable<int> Price { get; set; }
         public string ImagePro { get; set; }
         public Nullable<int> Quantity { get; set; }
-
+    
         public virtual Category Category1 { get; set; }
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<OrderDetail> OrderDetails { get; set; }
